feat: Renaming folders and including gitignore
We moved the project folder for pkg to a better readability and scoping other pkgs within. Since this are small packages, i changed to that instead of projects, which would be bigger stuff.
Also moved scripts folder to bin folder, to make it more distant to the commands we will need to run for python files.
Also included some files in gitignore to avoid a lot of unecessary staged files.